debug imu_fusion

+++ src/hedge_rcv_bin.cpp
@@ -431,7 +431,7 @@ int main(int argc, char **argv)
     if (hedgeIMUFusionReceiveCheck())
     {
 		ROS_INFO("IMU fusion: Address=%d, Timestamp: %08d, X=%.3f  Y= %.3f  Z=%.3f  q=%.3f,%.3f,%.3f,%.3f v=%.3f,%.3f,%.3f  a=%.3f,%.3f,%.3f",
-                (float) hedge_imu_fusion_msg.address,
+                (int) hedge_imu_fusion_msg.address,
 				(int) hedge_imu_fusion_msg.timestamp_ms,
 				(float) hedge_imu_fusion_msg.x_m, (float) hedge_imu_fusion_msg.y_m, (float) hedge_imu_fusion_msg.z_m,
 				(float) hedge_imu_fusion_msg.qw, (float) hedge_imu_fusion_msg.qx, (float) hedge_imu_fusion_msg.qy, (float) hedge_imu_fusion_msg.qz,
